simulator/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
opencv-python==4.9.0.80
numpy==1.26.3
pyyaml==6.0.1
python-multipart==0.0.6
